remove automated test generation. Remove integration test script
