+++ mlir/test/lit.cfg.py
@@ -59,7 +59,7 @@ llvm_config.with_system_environment(["HOME", "INCLUDE", "LIB", "TMP", "TEMP"])
 llvm_config.use_default_substitutions()
 
 # excludes: A list of directories and filenames to exclude from the testsuite.
-config.excludes = [".venv"]
+config.excludes = []
 
 # test_exec_root: The root path where tests should be run.
 config.test_exec_root = Path(config.project_binary_dir) / "test"
