+++ mlir/test/tools/qcc/qrisp-integration/README.md
@@ -1,21 +1,16 @@
 # From Qrisp to MLIR: Step-by-Step Guide
 
-This guide documents the process of converting a [Qrisp](https://qrisp.eu/) quantum program into an MLIR input file suitable for FullStaQD compilation (e.g. with `qcc`).
+This guide documents the process of converting a [Qrisp](https://qrisp.eu/) quantum program into an MLIR test case for `qcc`.
 
 ---
 
 ## Prerequisites
 
-This project requires a Python environment with `qrisp` installed. You can install it via pip:
+This project requires a Python environment set up by the `requirements.txt` in this folder. You can install it via pip:
 
 ```bash
-pip install git+https://github.com/eclipse-qrisp/Qrisp.git@b81ea2f979d21cd8d600e79d8b0c7066fe7cbe1b
-pip install xdsl==0.59.0
+pip install -r requirements.txt
 ```
-
-The above command is pinpointing to this a specific version of [Qrisp](https://github.com/eclipse-qrisp/Qrisp/pull/528) which mostly solved the dependencies from `stablehlo` dialect.
-
-## **Troubleshooting**: If you encounter any problems, check the official [Qrisp setup documentation](https://qrisp.eu/general/setup.html) for detailed instructions.
 
 ## Step 1 — Write Your Qrisp Program
 
@@ -41,28 +36,36 @@ This example creates a bell state state and measures all the qubits.
 Run the script:
 
 ```bash
-python your_script.py
-```
-
-The output will be a block of MLIR text. Copy it — this is your **raw MLIR**, which will require a few manual transformations before it is ready to use.
-
----
-
-## Step 2 — Add the `qcc.entry_point` Annotation
-
-Locate the main function header in the MLIR output — it will look roughly like:
-
-```mlir
-func.func @main(...) -> ... {
-```
-
-Add the `attributes { qcc.entry_point }` annotation immediately after the function signature, before the opening brace:
-
-```mlir
-func.func @main(...) -> ... attributes { qcc.entry_point } {
+python my_program.test.py > my_program.generated_test.mlir
 ```
 
 ---
+
+## Step 2 - Add lit RUN commands
+
+For example, these `RUN` commands enable checking against `QIR` and `qirrunner` simulation output:
+
+```
+// Test 1: Check QIR output.
+// RUN: qcc  %s | mlir-translate -mlir-to-llvmir | FileCheck %s --check-prefix=CHECK-QIR
+// Test 2: Check output recording from qir-runner simulation.
+// qir-runner requires a .ll file (it checks the extension and does not accept
+// stdin), so we first emit the LLVM IR to a temporary .ll file and then invoke
+// qir-runner on it.
+// RUN: qcc %s | mlir-translate -mlir-to-llvmir > %t.ll
+// RUN: uvx --from qirrunner qir-runner --file %t.ll -s 5 | FileCheck %s --check-prefix=CHECK-SIM
+```
+
+---
+
+## Step 3 - Add FileCheck directives
+
+Add checks for the different outputs:
+
+```
+// CHECK-QIR-DAG: ... for QIR output ...
+// CHECK-SIM:     ... for qirrunner output ...
+```
 
 ## Remark
