+++ mlir/test/tools/qcc/qrisp-integration/ghz.test.py
@@ -27,11 +27,8 @@ mlir = str(make_jaspr(prepare_ghz)().to_mlir(lower_stablehlo=True))
 
 print(
 f"""
-// RUN: not %python %S/integration_run.py -r 42 -n 1 -c qcc %s | FileCheck %s
+// GENERATED FROM QRISP.
 
 {mlir}
-
-// TODO: Once loop unrolling is in place, add proper checks
-// CHECK: qubit index must be a constant; unroll loops before this pass
 """
 )
